Move the implementation to a separate file
